+++ src/NetAppFiles/NetAppFiles/Models/PSNetAppFilesVolumeExportPolicy.cs
@@ -12,8 +12,6 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-using System.Collections.Generic;
-
 namespace Microsoft.Azure.Commands.NetAppFiles.Models
 {