+++ src/RecoveryServices/RecoveryServices.SiteRecovery/Models/PSFabricDetails.cs
@@ -1257,6 +1257,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
             this.RecoveryAzureCloudService = details.RecoveryCloudService;
             this.RecoveryFabricLocation = details.RecoveryFabricLocation;
             this.RecoveryAvailabilitySet = details.RecoveryAvailabilitySet;
+            this.RecoveryProximityPlacementGroupId = details.RecoveryProximityPlacementGroupId;
             this.TestFailoverRecoveryFabricObjectId = details.TestFailoverRecoveryFabricObjectId;
             this.MonitoringJobType = details.MonitoringJobType;
             this.MonitoringPercentageCompletion = details.MonitoringPercentageCompletion;
@@ -1361,6 +1362,11 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
         /// Recovery availability set.
         /// </summary>
         public string RecoveryAvailabilitySet { get; set; }
+
+        /// <summary>
+        /// Recovery proximity placement group Id.
+        /// </summary>
+        public string RecoveryProximityPlacementGroupId { get; set; }
 
         /// <summary>
         /// Synced configuration details of the virtual machine.